code reorganization: elf & bzimage modules
In preparation for more incoming code that's either elf or
bzimage specific, the respective blocks of functionality
now sit in separate modules. This makes the code easier to
navigate and improves readability.

Fixes #22 as well.
