+++ src/loader/elf/test_badnote.bin
